- convertion: time to string fix, int to string hex fix  - logging: datetime fix; init refactoring  - threading: worker refactoring  - networking: freeze sockets implementation - not work; dns for windows; wsa wrapper for windows  - parsing: node to string fix; serializator bugfix  - platform macroses refactoring  - changed project file structure  - vs project files

+++ source/utility/fs/implementation/fast_text_writer.cpp
@@ -57,8 +57,8 @@ void CFastTextWriter::writeLine(string const &line_) {
     if (!m_file)
         throw std::runtime_error("fast text writer not initialized"); // ----->
 
-    auto        line    = line_ + "\n";
-    uint32_t    count   = fwriteInternal(line.c_str(), 1, line.size(), m_file);
+    auto    line    = line_ + "\n";
+    size_t	count   = fwriteInternal(line.c_str(), 1, line.size(), m_file);
 
     if (count != line.size())
         throw std::runtime_error(
